split pre.py to subfiles
