+++ rom_operator_inference/pre/__init__.py
@@ -1,0 +1,7 @@
+# pre/__init__.py
+"""Tools for preprocessing data."""
+
+from ._basis import *
+from ._shift_scale import *
+from ._reprojection import *
+from ._finite_differences import *
